+++ cider-client.el
@@ -28,6 +28,7 @@
 (require 'map)
 (require 'seq)
 (require 'subr-x)
+(require 'parseedn)
 
 (require 'clojure-mode)
 (require 'spinner)
@@ -412,12 +413,36 @@ contain a `candidates' key, it is returned as is."
           info)
       var-info)))
 
+(defconst cider-info-form "
+(do
+  (require 'clojure.java.io)
+  (require 'clojure.walk)
+
+  (if-let [var (resolve '%s)]
+    (let [info (meta var)]
+      (-> info
+          (update :ns str)
+          (update :name str)
+          (update :file (comp str clojure.java.io/resource))
+          (assoc :arglists-str (str (:arglists info)))
+          (clojure.walk/stringify-keys)))))
+")
+
+(defun cider-fallback-eval:info (var)
+  "Obtain VAR metadata via a regular eval.
+Used only when the info nREPL middleware is not available."
+  (let* ((response (cider-sync-tooling-eval (format cider-info-form var)))
+         (var-info (nrepl-dict-from-hash (parseedn-read-str (nrepl-dict-get response "value")))))
+    var-info))
+
 (defun cider-var-info (var &optional all)
   "Return VAR's info as an alist with list cdrs.
 When multiple matching vars are returned you'll be prompted to select one,
 unless ALL is truthy."
   (when (and var (not (string= var "")))
-    (let ((var-info (cider-sync-request:info var)))
+    (let ((var-info (if (nrepl-op-supported-p "info")
+                        (cider-sync-request:info var)
+                      (cider-fallback-eval:info var))))
       (if all var-info (cider--var-choice var-info)))))
 
 (defun cider-member-info (class member)